requirements.txt
alembic==0.9.8
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.2.1
Django==1.11.10
dominate==2.3.1
ez-setup==0.9
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Images==2.1.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
image==1.5.19
infinity==1.4
intervals==0.8.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
MySQL-python==1.2.5
pathlib==1.0.1
# PIL==1.1.7
Pillow==5.0.0
pillowcase==2.0.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2018.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.4
SQLAlchemy-Utils==0.33.1
stripe==1.79.1
urllib3==1.22
validators==0.12.1
virtualenv==15.1.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
WTForms-Alchemy==0.16.6
WTForms-Components==0.10.3
